+++ src/reader/plugins/mark_as_read.py
@@ -1,11 +1,11 @@
 """
-reader.mark_as_read
-~~~~~~~~~~~~~~~~~~~
+reader.mark_entry_as_read
+~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Mark added entries of specific feeds as read if their title matches a regex.
 
 To configure, set the ``make_reader_reserved_name('mark_as_read')``
-(by default, ``.reader.mark_as_read``)
+(by default, ``.reader.mark_entry_as_read``)
 feed metadata to something like::
 
     {
@@ -58,7 +58,7 @@ def _mark_as_read(reader, entry):
 
     for pattern in patterns or ():
         if re.search(pattern, entry.title):
-            reader.mark_as_read(entry)
+            reader.mark_entry_as_read(entry)
             return